Rename engine to line_editor
